Milestone
Stopping at a point where in the class Guts we are starting to
work through the AST to get all the documentation bits we need.

I was going to model after docgen but have decided to switching to
building up an entire data model that hangs off of a Bundle then
once that is fully constructed, hand that off to a backend for
final processing.

This will mean that whatever "Guts" is, won't need to know anything
about Backend. Guts returns a Bundle when done and then that
can be handed off to the Backend for processing.