Push public Docker images to GitHub Container Registry (#9)
We are transitioning from DockerHub to GitHub Container Registry. As
part of this transition, we will have a period of time where we push
all images to both registries.

This commit updates ponydoc to push its public images of latest and
release builds to GitHub Container Registry and DockerHub.

In addition, it adds a job to run after the latest docker image is
pushed to prune old "latest" images that have become "untagged" and
shouldn't be being used any longer.